Change instructions parser from static to non-static

+++ MartianRobots/BusinessObjects/Instruction.cs
@@ -8,10 +8,10 @@ namespace MartianRobots.BusinessObjects
     public class Instruction
     {
         public string Name { get; set; }
-        public double Degrees { get; set; }
+        public int Degrees { get; set; }
         public int Distance { get; set; }
 
-        public Instruction(string name, int distance, double degrees = 0)
+        public Instruction(string name, int distance, int degrees = 0)
         {
             Name = name;
             Degrees = degrees;
